recursion_fibonacci updated to be Mathematica like
Previous implementation of 'recursion_fibonacci' method was C code in Mathematica. Changed it to Mathematica-like implementation wrapped in Module[] to avoid cashing. On my Macbook pro the improvement is 42 times... 

It is even faster if Module[] isolation is not used, but Mathematica does caching of previous results thus it would be not a fair comparison. Read more on: http://community.wolfram.com/groups/-/m/t/442573